Update Spinner Component and add documentation
